Rotation and chances
Added a chance for another effect. Now it can rotates 90 in both sides. To make it work with Random Thoughts Generation, I added a chance of encounter for either event to happen, which can be easily manipulated in their SO. Also, if their combined percentage is less than 100, then the missing percent goes for another effect, which is to do... nothing. In other words, it goes to the node and immediately to the next one.